Moves threading/synchronization to `iree/base/threading/`. Elevates threading primitives from scattered internal utilities to a proper first-class module. The threading code was previously buried in `iree/base/internal/` alongside unrelated utilities and was not exported for public use. This reorganization makes the threading API surface explicit and sets up a foundation for future NUMA and concurrency work that will expand the file count.
New/moved module components:
- `api.h`: Umbrella header for convenient inclusion
- `mutex.h`: Fat mutex and slim mutex with detailed platform documentation
- `notification.h`: Wait-free event counters for lock-free coordination
- `thread.h`: Thread creation, affinity, and lifecycle management
- `call_once.h`: One-time initialization primitive (moved from internal)

Trampoline headers were left in iree/base/internal/ for future cleanup.
