+++ runtime/bindings/tflite/interpreter.c
@@ -6,7 +6,7 @@
 
 #include "runtime/bindings/tflite/interpreter.h"
 
-#include "iree/base/internal/call_once.h"
+#include "iree/base/threading/call_once.h"
 #include "iree/hal/drivers/init.h"
 #include "iree/modules/hal/module.h"
 #include "runtime/bindings/tflite/model.h"
