+++ runtime/src/iree/base/internal/wait_handle_inproc.c
@@ -14,9 +14,9 @@
 #include <string.h>
 
 #include "iree/base/api.h"
-#include "iree/base/internal/call_once.h"
 #include "iree/base/internal/synchronization.h"
 #include "iree/base/internal/wait_handle.h"
+#include "iree/base/threading/call_once.h"
 
 // This implementation uses iree_notification_t - backed by a futex in most
 // cases - to simulate system wait handles. When using a single handle such as
