+++ runtime/src/iree/base/threading/BUILD.bazel
@@ -1,0 +1,84 @@
+# Copyright 2026 The IREE Authors
+#
+# Licensed under the Apache License v2.0 with LLVM Exceptions.
+# See https://llvm.org/LICENSE.txt for license information.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+load("//build_tools/bazel:build_defs.oss.bzl", "iree_runtime_cc_library", "iree_runtime_cc_test")
+
+package(
+    default_visibility = ["//visibility:public"],
+    features = ["layering_check"],
+    licenses = ["notice"],  # Apache 2.0
+)
+
+iree_runtime_cc_library(
+    name = "threading",
+    srcs = [
+        "mutex.c",
+        "notification.c",
+        "numa_fallback.c",
+        "numa_impl.h",
+        "numa_linux.c",
+        "numa_win32.c",
+    ],
+    hdrs = [
+        "api.h",
+        "call_once.h",
+        "futex.h",
+        "mutex.h",
+        "notification.h",
+        "numa.h",
+    ],
+    deps = [
+        "//build_tools:pthreads",
+        "//runtime/src/iree/base",
+        "//runtime/src/iree/base:core_headers",
+        "//runtime/src/iree/base/internal",
+        "//runtime/src/iree/base/internal:memory",
+    ],
+)
+
+iree_runtime_cc_test(
+    name = "futex_test",
+    srcs = ["futex_test.cc"],
+    deps = [
+        ":threading",
+        "//runtime/src/iree/base",
+        "//runtime/src/iree/base/internal:threading",
+        "//runtime/src/iree/testing:gtest",
+        "//runtime/src/iree/testing:gtest_main",
+    ],
+)
+
+iree_runtime_cc_test(
+    name = "numa_test",
+    srcs = ["numa_test.cc"],
+    deps = [
+        ":threading",
+        "//runtime/src/iree/testing:gtest",
+        "//runtime/src/iree/testing:gtest_main",
+    ],
+)
+
+iree_runtime_cc_library(
+    name = "thread",
+    srcs = [
+        "thread.c",
+        "thread_darwin.c",
+        "thread_impl.h",
+        "thread_pthreads.c",
+        "thread_win32.c",
+    ],
+    hdrs = [
+        "thread.h",
+    ],
+    deps = [
+        ":threading",
+        "//build_tools:dl",
+        "//build_tools:pthreads",
+        "//runtime/src/iree/base",
+        "//runtime/src/iree/base:core_headers",
+        "//runtime/src/iree/base/internal",
+    ],
+)
