+++ runtime/src/iree/base/threading/CMakeLists.txt
@@ -1,0 +1,82 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# runtime/src/iree/base/threading/BUILD.bazel                                  #
+#                                                                              #
+# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
+# CMake-only content.                                                          #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    threading
+  HDRS
+    "api.h"
+    "call_once.h"
+    "futex.h"
+    "mutex.h"
+    "notification.h"
+    "numa.h"
+  SRCS
+    "mutex.c"
+    "notification.c"
+    "numa_fallback.c"
+    "numa_impl.h"
+    "numa_linux.c"
+    "numa_win32.c"
+  DEPS
+    iree::base
+    iree::base::core_headers
+    iree::base::internal
+    iree::base::internal::memory
+  PUBLIC
+)
+
+iree_cc_test(
+  NAME
+    futex_test
+  SRCS
+    "futex_test.cc"
+  DEPS
+    ::threading
+    iree::base
+    iree::base::internal::threading
+    iree::testing::gtest
+    iree::testing::gtest_main
+)
+
+iree_cc_test(
+  NAME
+    numa_test
+  SRCS
+    "numa_test.cc"
+  DEPS
+    ::threading
+    iree::testing::gtest
+    iree::testing::gtest_main
+)
+
+iree_cc_library(
+  NAME
+    thread
+  HDRS
+    "thread.h"
+  SRCS
+    "thread.c"
+    "thread_darwin.c"
+    "thread_impl.h"
+    "thread_pthreads.c"
+    "thread_win32.c"
+  DEPS
+    ${CMAKE_DL_LIBS}
+    ::threading
+    iree::base
+    iree::base::core_headers
+    iree::base::internal
+  PUBLIC
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
