+++ runtime/src/iree/base/threading/api.h
@@ -1,0 +1,17 @@
+// Copyright 2026 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#ifndef IREE_BASE_THREADING_API_H_
+#define IREE_BASE_THREADING_API_H_
+
+#include "iree/base/threading/call_once.h"
+#include "iree/base/threading/futex.h"
+#include "iree/base/threading/mutex.h"
+#include "iree/base/threading/notification.h"
+#include "iree/base/threading/numa.h"
+#include "iree/base/threading/thread.h"
+
+#endif  // IREE_BASE_THREADING_API_H_
