+++ runtime/src/iree/base/threading/thread_pthreads.c
@@ -5,7 +5,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 // NOTE: must be first to ensure that we can define settings for all includes.
-#include "iree/base/internal/threading_impl.h"
+#include "iree/base/threading/thread_impl.h"
 
 #if defined(IREE_PLATFORM_ANDROID) || defined(IREE_PLATFORM_EMSCRIPTEN) || \
     defined(IREE_PLATFORM_LINUX)
@@ -21,9 +21,9 @@
 #include <unistd.h>
 
 #include "iree/base/internal/atomics.h"
-#include "iree/base/internal/call_once.h"
-#include "iree/base/internal/synchronization.h"
-#include "iree/base/internal/threading.h"
+#include "iree/base/threading/call_once.h"
+#include "iree/base/threading/notification.h"
+#include "iree/base/threading/thread.h"
 
 #if defined(IREE_PLATFORM_EMSCRIPTEN)
 #include <emscripten/threading.h>
