+++ runtime/src/iree/builtins/ukernel/tools/BUILD.bazel
@@ -22,6 +22,7 @@ iree_runtime_cc_library(
         "//runtime/src/iree/base/internal",
         "//runtime/src/iree/base/internal:cpu",
         "//runtime/src/iree/base/internal:synchronization",
+        "//runtime/src/iree/base/threading",
         "//runtime/src/iree/builtins/ukernel",
         "//runtime/src/iree/schemas:cpu_data",
     ],
