+++ runtime/src/iree/builtins/ukernel/tools/CMakeLists.txt
@@ -22,6 +22,7 @@ iree_cc_library(
     iree::base::internal
     iree::base::internal::cpu
     iree::base::internal::synchronization
+    iree::base::threading
     iree::builtins::ukernel
     iree::schemas::cpu_data
   PUBLIC
