+++ runtime/src/iree/hal/BUILD.bazel
@@ -76,6 +76,7 @@ iree_runtime_cc_library(
         "//runtime/src/iree/base/internal",
         "//runtime/src/iree/base/internal:path",
         "//runtime/src/iree/base/internal:synchronization",
+        "//runtime/src/iree/base/threading",
         "//runtime/src/iree/io:file_handle",
     ],
 )
