+++ runtime/src/iree/hal/driver_registry.c
@@ -9,9 +9,9 @@
 #include <stddef.h>
 #include <string.h>
 
-#include "iree/base/internal/call_once.h"
 #include "iree/base/internal/path.h"
 #include "iree/base/internal/synchronization.h"
+#include "iree/base/threading/call_once.h"
 
 //===----------------------------------------------------------------------===//
 // iree_hal_driver_registry_t
