+++ runtime/src/iree/tooling/CMakeLists.txt
@@ -112,6 +112,7 @@ iree_cc_library(
     iree::base
     iree::base::internal::flags
     iree::base::internal::synchronization
+    iree::base::threading
     iree::hal
     iree::hal::drivers
     iree::hal::utils::allocators
